+++ src/components/shared/basic/ReleaseProcess/OfferCard/index.tsx
@@ -542,7 +542,7 @@ export default function OfferCard() {
         isValid={isValid}
         loader={loading}
         helpUrl={
-          '/documentation/?path=user%2F05.+Service%28s%29%2F02.+Service+Release+Process'
+          '/documentation/?path=user%2F05.+Service%28s%29%2F02.+Service+Release+Process%2F01.+Service+Card.md'
         }
       />
     </div>